Remove unnecessary effect hooks
State hooks was enought to handle the fact to render some changes
the DOM. useEffect had no reason to implemented in Post.js.
However, Wall.js required to getPostList in effect hook to refresh
the entire list od posts.

As a result, the performance of the app is improved since effect
hooks are not used unnecessarllyE