Add Stage environments in React
Building to prodution required to change the API path manuelly, which
led to a non-productive repetitive task.

Stage evironment variables came into acction. 'dev' and 'prod' were
the two stages included so far. Config.js has the variable for each
environment.

As a result, automated selection of envorinment variables when
passsing the desired stage to REACT_APP_STAGE variable to star or
build the app.

ReactStageEnvironmentVariable